Converted all containers but the maps and examples to signed sizes and indices.

+++ misc/examples/new_pque.c
@@ -48,7 +48,7 @@ int main()
         cpque_int_push(&ique, 123);
         cpque_int_push(&ique, 321);
         // print
-        for (size_t i=0; i<cpque_int_size(&ique); ++i)
+        for (int i=0; i<cpque_int_size(&ique); ++i)
             printf(" %d", ique.data[i]);
         puts("");
     }
